refactor(watt): extract segmented buttons showcase to its own file
Move the WattStorybookSegmentedButtonsShowcaseComponent into
storybook-segmented-buttons-showcase.component.ts, matching the
storybook-*.component.ts pattern used elsewhere in Watt (e.g. drawer,
colors).
